Remove unnecessary cast in convertColor()

+++ jsvg-javafx/src/main/java/com/github/weisj/jsvg/renderer/jfx/impl/FXAWTBridge.java
@@ -203,7 +203,7 @@ public class FXAWTBridge {
                 awtColor.getRed(),
                 awtColor.getGreen(),
                 awtColor.getBlue(),
-                ((double) awtColor.getAlpha() / 255.0) * awtGlobalOpacity);
+                (awtColor.getAlpha() / 255.0) * awtGlobalOpacity);
     }
 
     public static javafx.scene.paint.Color convertRGBColor(@NotNull RGBColor rgbColor, double awtGlobalOpacity) {
